% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_colors.R
\name{display_colors}
\alias{display_colors}
\title{display_colors}
\usage{
display_colors(Name)
}
\arguments{
\item{Name}{character names, you can see all names using get_character_name}
}
\value{
a color plot
}
\description{
display the colors of character
}
\examples{
display_colors('ZHONGLI')
}
