% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genshinpalette.R
\name{Genshinpalette}
\alias{Genshinpalette}
\title{Genshinpalette}
\usage{
Genshinpalette(Name, Count = 6)
}
\arguments{
\item{Name}{character names, you can see all names using get_character_name}

\item{Count}{the number of colors you want to generate}
}
\value{
a vector of colors
}
\description{
Genshinpalette Generate a color palette based on the theme
color of Genshin character creation
}
\examples{
Genshinpalette('ZHONGLI',8)

}
