% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_character_name.R
\name{get_character_name}
\alias{get_character_name}
\title{get_character_name}
\usage{
get_character_name(Country = "All")
}
\arguments{
\item{Country}{a country name could be one of All, Mondstadt, Liyue, Inazuma, Sumeru, Fontaine, Natlan, Snezhnaya}
}
\value{
a vector that contains character name, you can use these name to Genshinpalette
}
\description{
get_character_name can obtain the character name in Teyvat
}
\examples{
get_character_name('Liyue')

}
