% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_Genshin.R
\name{scale_color_Genshin}
\alias{scale_color_Genshin}
\title{Genshin Color Scales}
\description{Genshin Color Scales
including
\code{scale_color_Genshin()}
\code{scale_colour_Genshin()}
\code{scale_fill_Genshin()}}
\usage{
scale_color_Genshin(
  palette = names(internal_da$countrylist$All),
  alpha = 1,
  ...
)
}
\arguments{
\item{palette}{character name, same as Name in Genshinpalette()}

\item{alpha}{default 1}

\item{...}{additional parameters for \code{\link[ggplot2]{discrete_scale}}}
}
\description{
Genshin Color Scales
including
\code{scale_color_Genshin()}
\code{scale_colour_Genshin()}
\code{scale_fill_Genshin()}
}
\examples{
library("ggplot2")
data("diamonds")

ggplot(
  subset(diamonds, carat >= 2.2),
  aes(x = table, y = price, colour = cut)
) +
  geom_point(alpha = 0.7) +
  geom_smooth(method = "loess", alpha = 0.1, size = 1, span = 1) +
  theme_bw() + scale_color_Genshin('ZHONGLI')

ggplot(
  subset(diamonds, carat > 2.2 & depth > 55 & depth < 70),
  aes(x = depth, fill = cut)
) +
  geom_histogram(colour = "black", binwidth = 1, position = "dodge") +
  theme_bw() + scale_fill_Genshin('ZHONGLI')
}
